refactor(blockchain): move attestation coverage emission into its own module
Address review feedback: relocate all attestation aggregate coverage
observability into a new crates/blockchain/src/coverage.rs. The pre-merge
new_payloads snapshot is now captured by the actor in BlockChainServer::on_tick
(when proposing or at the end of the slot) instead of being threaded as a
parameter through store::on_tick / accept_new_attestations /
get_proposal_head / produce_block_with_signatures. Since the snapshot is owned
solely by the actor and only touched from its single-threaded message loop,
the Arc<Mutex> handle is replaced by a plain Option<CoverageSnapshot> field.
No behavior change - coverage emission is observability-only.
