+++ crates/blockchain/tests/signature_spectests.rs
@@ -50,7 +50,7 @@ fn run(path: &Path) -> datatest_stable::Result<()> {
 
         // Advance time to the block's slot
         let block_time_ms = genesis_time * 1000 + signed_block.message.slot * MILLISECONDS_PER_SLOT;
-        store::on_tick(&mut st, block_time_ms, true, None);
+        store::on_tick(&mut st, block_time_ms, true);
 
         // Process the block (this includes signature verification)
         let result = store::on_block(&mut st, signed_block);
